requirements.txt
alembic==1.17.2
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==21.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
psutil==7.1.3
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
Werkzeug==3.1.3
WTForms==3.2.1
openpyxl==3.1.2
